Rename packages to adhere to convention
