+++ src/main/java/duke/command/DoneCommand.java
@@ -1,9 +1,9 @@
-package duke.commands;
+package duke.command;
 
-import duke.exceptions.BadInputFormatException;
-import duke.exceptions.NoSuchTaskException;
-import duke.utils.TaskList;
-import duke.utils.Ui;
+import duke.exception.BadInputFormatException;
+import duke.exception.NoSuchTaskException;
+import duke.util.TaskList;
+import duke.util.Ui;
 
 public class DoneCommand extends Command {
     private int index;
