+++ src/main/java/duke/command/ExitCommand.java
@@ -1,8 +1,8 @@
-package duke.commands;
+package duke.command;
 
-import duke.exceptions.BadInputFormatException;
-import duke.utils.TaskList;
-import duke.utils.Ui;
+import duke.exception.BadInputFormatException;
+import duke.util.TaskList;
+import duke.util.Ui;
 
 import java.io.IOException;
