+++ src/main/java/duke/exception/BadInputFormatException.java
@@ -1,4 +1,4 @@
-package duke.exceptions;
+package duke.exception;
 
 public class BadInputFormatException extends InvalidInputException {
     public BadInputFormatException() {
