+++ src/main/java/duke/task/Task.java
@@ -1,4 +1,4 @@
-package duke.tasks;
+package duke.task;
 
 /**
  * Represents a task object.
